pyproject.toml
[project]
name = "zerde-bot"
version = "0.1.0"
description = "A Serverless Telegram Bot for IT Community Management, built with Python and AWS CDK."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Bayashat", email = "bayashat.tokmukamet@gmail.com" }
]

keywords = [
    "telegram-bot",
    "aws-cdk",
    "serverless",
    "lambda",
    "community-management",
    "python3.13"
]
dependencies = [
    "aws-cdk-aws-lambda-python-alpha>=2.248.0a0",
    "aws-cdk-lib>=2.236.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "aws-lambda-powertools>=3.24.0",
    "boto3>=1.42.38",
    "feedparser>=6.0.12",
    "google-genai>=1.65.0",
    "pre-commit>=4.5.1",
    "requests>=2.32.5",
]
